+++ src/__tests__/componentsTest/TeamCard.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import TeamCard from "../../components/Cards/TeamCard";
+import TeamCard from "../../components/Card/TeamCard";
 
 describe("TeamCard Component", () => {
   const teamMembers = [
